requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Reuploaded==1.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.13.2
Werkzeug==3.1.3
